// src/types/aboutPageContent.type.ts
export type AboutPageContentType = {
  aboutInformation: string;
  ourVision: string;
  ourMission: string;
  ourPurpose: string;
  areasOfFocus: string;
};

export type AboutPageContentHiddenFieldsType = {
  aboutInformation: boolean;
  ourVision: boolean;
  ourMission: boolean;
  ourPurpose: boolean;
  areasOfFocus: boolean;
};